WIP expressions

+++ biscuit-datalog/src/matched_variables.zig
@@ -29,10 +29,10 @@ const Term = @import("term.zig").Term;
 /// }
 /// ```
 pub const MatchedVariables = struct {
-    variables: std.AutoHashMap(u64, ?Term),
+    variables: std.AutoHashMap(u32, ?Term),
 
     pub fn init(allocator: mem.Allocator, rule: *Rule) !MatchedVariables {
-        var variables = std.AutoHashMap(u64, ?Term).init(allocator);
+        var variables = std.AutoHashMap(u32, ?Term).init(allocator);
 
         // Add all variables in predicates in the rule's body to variable set
         for (rule.body.items) |predicate| {
@@ -56,7 +56,7 @@ pub const MatchedVariables = struct {
         return .{ .variables = variables };
     }
 
-    pub fn get(matched_variables: *const MatchedVariables, key: u64) ?Term {
+    pub fn get(matched_variables: *const MatchedVariables, key: u32) ?Term {
         return matched_variables.variables.get(key) orelse return null;
     }
 
@@ -66,7 +66,7 @@ pub const MatchedVariables = struct {
     ///
     /// If the variable is unset we bind to the term unconditionally and
     /// return true.
-    pub fn insert(matched_variables: *MatchedVariables, variable: u64, term: Term) !bool {
+    pub fn insert(matched_variables: *MatchedVariables, variable: u32, term: Term) !bool {
         const entry = matched_variables.variables.getEntry(variable) orelse return false;
 
         if (entry.value_ptr.*) |existing_term| {
@@ -94,15 +94,15 @@ pub const MatchedVariables = struct {
 
     /// If every variable in MatchedVariables has been assigned a term return a map
     /// from variable -> non-null term, otherwise return null.
-    pub fn complete(matched_variables: *const MatchedVariables, allocator: mem.Allocator) !?std.AutoHashMap(u64, Term) {
+    pub fn complete(matched_variables: *const MatchedVariables, allocator: mem.Allocator) !?std.AutoHashMap(u32, Term) {
         if (!matched_variables.isComplete()) return null;
 
-        var completed_variables = std.AutoHashMap(u64, Term).init(allocator);
+        var completed_variables = std.AutoHashMap(u32, Term).init(allocator);
         errdefer completed_variables.deinit();
 
         var it = matched_variables.variables.iterator();
         while (it.next()) |kv| {
-            const key: u64 = kv.key_ptr.*;
+            const key: u32 = kv.key_ptr.*;
             const value: ?Term = kv.value_ptr.*;
 
             try completed_variables.put(key, value.?);
